feat: finish playlist module

+++ src/playlists/playlist.service.ts
@@ -1,44 +1,62 @@
-import { Injectable, NotFoundException} from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException} from '@nestjs/common';
 import { CreatePlaylistDto } from './dto/create-playlist.dto';
 import { UpdatePlaylistDto } from './dto/update-playlist.dto';
 import { PlaylistRepository } from './playlist.repository';
-import { UserId } from 'src/auth/decorators/userId.decorator';
 import { S3Service } from 'src/files/services/s3.service';
-import { ManagedUpload } from 'aws-sdk/clients/s3';
 import { MusicsRepository } from 'src/musics/musics.repository';
-import { count } from 'console';
+import { UsersRepository } from 'src/users/users.repository';
+
 
 
 @Injectable()
 export class PlaylistService {
   constructor ( private readonly playlistRepository: PlaylistRepository,
                 private readonly s3Service: S3Service,
+                private readonly usersRepository: UsersRepository,
                 private musicsRepository: MusicsRepository,
    ) {}
 
 
-  async create(createPlaylistDto: CreatePlaylistDto, file: Express.Multer.File) {
-    let image: ManagedUpload.SendData = null
-    if(file){
-      
-      image = await this.s3Service.upload(file)
+   async create(createPlaylistDto: CreatePlaylistDto, file: Express.Multer.File, userId: number) {
+    if (!file) {
+      throw new NotFoundException('File is required for creating a playlist');
     }
   
-    return  await this.playlistRepository.create(createPlaylistDto, image);
+    const user = await this.usersRepository.findById(userId);
+    if (!user) {
+      throw new NotFoundException(`User with ID ${userId} not found`);
+    }
+  
+    const image = await this.s3Service.upload(file);
+
+    return await this.playlistRepository.create(createPlaylistDto, image, user);
   }
 
 
   async addMusicToPlaylist(playlistId:number,musicId:number) {
-    const playlist = await this.playlistRepository.findOne(playlistId)
+    const playlist = await this.playlistRepository.findOne(playlistId) 
+    
+    console.log(playlist)
+    
     if(!playlist) throw new NotFoundException(`playlist with given id ${playlistId} not found`)
 
     const music = await this.musicsRepository.findOne(musicId)
     if(!music) throw new NotFoundException(`music with id ${musicId} not found`)
-    // playlist.count = count++
-    // playlist.music.push(music)
-    // playlist.count = playlist.music.length
+    
 
-    // return this.playlistRepository.save(playlist)
+    const musicExist = playlist.music.some(m => m.id === music.id) 
+    if(musicExist){
+      throw new BadRequestException('Music already exist in playlist');
+    }
+    
+    playlist.music.push(music)
+   
+
+   playlist.count++
+
+   console.log(playlist.count)
+   
+    return this.playlistRepository.save(playlist)
 }
 
   async findAll(userId:number) {
@@ -55,7 +73,12 @@ export class PlaylistService {
   }
 
   async delete(id: number) {
-    return await this.playlistRepository.remove(id);
+    const playlist = await this.playlistRepository.findOne(id)
+
+    if(!playlist) {
+      throw new NotFoundException(`Playlist with ID ${id} not found`)
+    }
+    return await this.playlistRepository.remove(id)
   }
